requirements.txt
cx-Freeze==6.8.1
cx-Logging==3.0
importlib-metadata==4.8.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pytesseract==0.3.8
zipp==3.6.0
